Fixing termination detection of scheduling loop
 - Partially reverting to old detection mechanism
 - Only cleaning up currently running thread
 - Letting background thread phase out correctly

+++ hpx/runtime/threads/policies/local_priority_queue_scheduler.hpp
@@ -429,8 +429,8 @@ namespace hpx { namespace threads { namespace policies
         bool cleanup_terminated(std::size_t num_thread, bool delete_all)
         {
             bool empty = true;
-//             if (num_thread == std::size_t(-1))
-//             {
+            if (num_thread == std::size_t(-1))
+            {
                 for (std::size_t i = 0; i != queues_.size(); ++i)
                     empty = queues_[i]->cleanup_terminated(delete_all) && empty;
                 if (!delete_all)
@@ -441,21 +441,20 @@ namespace hpx { namespace threads { namespace policies
                         cleanup_terminated(delete_all) && empty;
 
                 empty = low_priority_queue_.cleanup_terminated(delete_all) && empty;
-                return empty;
-//             }
-//             else
-//             {
-//                 empty = queues_[num_thread]->cleanup_terminated(delete_all);
-//                 if (delete_all)
-//                     return true;
-//
-//                 if (num_thread < high_priority_queues_.size())
-//                     empty = high_priority_queues_[num_thread]->
-//                         cleanup_terminated(delete_all) && empty;
-//
-//                 low_priority_queue_.cleanup_terminated(delete_all);
-//                 return true;
-//             }
+            }
+            else
+            {
+                empty = queues_[num_thread]->cleanup_terminated(delete_all);
+                if (delete_all)
+                    return true;
+
+                if (num_thread < high_priority_queues_.size())
+                    empty = high_priority_queues_[num_thread]->
+                        cleanup_terminated(delete_all) && empty;
+
+                empty = low_priority_queue_.cleanup_terminated(delete_all) && empty;
+            }
+            return empty;
         }
 
         ///////////////////////////////////////////////////////////////////////
@@ -498,18 +497,15 @@ namespace hpx { namespace threads { namespace policies
 
             // now create the thread
             if (data.priority == thread_priority_high_recursive ||
-                data.priority == thread_priority_high)
+                data.priority == thread_priority_high ||
+                data.priority == thread_priority_boost)
             {
+                if (data.priority == thread_priority_boost)
+                {
+                    data.priority = thread_priority_normal;
+                }
                 std::size_t num = num_thread % high_priority_queues_.size();
-                high_priority_queues_[num]->create_thread(data, id,
-                    initial_state, run_now, ec);
-                return;
-            }
 
-            if (data.priority == thread_priority_boost)
-            {
-                data.priority = thread_priority_normal;
-                std::size_t num = num_thread % high_priority_queues_.size();
                 high_priority_queues_[num]->create_thread(data, id,
                     initial_state, run_now, ec);
                 return;
@@ -604,19 +600,6 @@ namespace hpx { namespace threads { namespace policies
             if (std::size_t(-1) == num_thread)
                 num_thread = curr_queue_++ % queue_size;
 
-            // Select a OS thread which hasn't been disabled
-            auto const& rp = resource::get_partitioner();
-            while (true)
-            {
-                auto mask = rp.get_pu_mask(
-                    num_thread + parent_pool_->get_thread_offset());
-
-                if (bit_and(mask, parent_pool_->get_used_processing_units()))
-                    break;
-
-                num_thread = (num_thread + 1) % queue_size;
-            }
-
             if (priority == thread_priority_high_recursive ||
                 priority == thread_priority_high ||
                 priority == thread_priority_boost)
@@ -642,19 +625,6 @@ namespace hpx { namespace threads { namespace policies
             std::size_t queue_size = queues_.size();
             if (std::size_t(-1) == num_thread)
                 num_thread = curr_queue_++ % queue_size;
-
-            // Select a OS thread which hasn't been disabled
-            auto const& rp = resource::get_partitioner();
-            while (true)
-            {
-                auto mask = rp.get_pu_mask(
-                    num_thread + parent_pool_->get_thread_offset());
-
-                if (bit_and(mask, parent_pool_->get_used_processing_units()))
-                    break;
-
-                num_thread = (num_thread + 1) % queue_size;
-            }
 
             if (priority == thread_priority_high_recursive ||
                 priority == thread_priority_high ||
@@ -959,17 +929,6 @@ namespace hpx { namespace threads { namespace policies
         virtual bool wait_or_add_new(std::size_t num_thread, bool running,
             std::int64_t& idle_loop_count)
         {
-            // Check if we have been disabled
-            {
-                auto const& rp = resource::get_partitioner();
-                auto mask = rp.get_pu_mask(
-                    num_thread + parent_pool_->get_thread_offset());
-
-                if (!bit_and(mask, parent_pool_->get_used_processing_units()))
-                    return true;
-
-            }
-
             std::size_t added = 0;
             bool result = true;
 
@@ -989,6 +948,18 @@ namespace hpx { namespace threads { namespace policies
             result = this_queue->wait_or_add_new(
                 running, idle_loop_count, added) && result;
             if (0 != added) return result;
+
+            // Check if we have been disabled
+            {
+                auto const& rp = resource::get_partitioner();
+                auto mask = rp.get_pu_mask(
+                    num_thread + parent_pool_->get_thread_offset());
+
+                if (!bit_and(mask, parent_pool_->get_used_processing_units()))
+                {
+                    return added == 0 && !running;
+                }
+            }
 
             for (std::size_t idx: victim_threads_[num_thread])
             {
@@ -1051,7 +1022,6 @@ namespace hpx { namespace threads { namespace policies
 
             result = low_priority_queue_.wait_or_add_new(running,
                 idle_loop_count, added) && result;
-            if (0 != added) return result;
 
             return result;
         }
